requirements.txt
fastapi
line-bot-sdk
uvicorn
python-dotenv
google-cloud-secret-manager